Renamed directory for extra classes needed for tests for task 4 (hw2)
